refactor(tools): extract shell request prep helpers
